Extract hostgroup to separete module (for 'host' feature)
