+++ src/host/create.rs
@@ -2,21 +2,12 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-use crate::{r#macro::model::ZabbixHostMacro, template::model::ZabbixTemplate};
+use crate::{
+    hostgroup::model::ZabbixHostGroupId, r#macro::model::ZabbixHostMacro,
+    template::model::ZabbixTemplate,
+};
 
-use super::model::{ZabbixHostGroupId, ZabbixHostInterface, ZabbixHostTag};
-
-/// API: https://www.zabbix.com/documentation/6.0/en/manual/api/reference/hostgroup/create
-#[derive(Serialize, Debug)]
-pub struct CreateHostGroupRequest {
-    pub name: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct CreateHostGroupResponse {
-    #[serde(rename = "groupids")]
-    pub group_ids: Vec<String>,
-}
+use super::model::{ZabbixHostInterface, ZabbixHostTag};
 
 /// API: https://www.zabbix.com/documentation/6.0/en/manual/api/reference/host/create
 #[derive(Serialize, Debug)]
