+++ src/hostgroup/mod.rs
@@ -1,0 +1,3 @@
+pub mod create;
+pub mod get;
+pub mod model;
